+++ sdk/src/relay/api/mod.rs
@@ -1,5 +1,7 @@
+mod stream_events;
 mod subscribe;
 mod try_connect;
 
+pub use self::stream_events::*;
 pub use self::subscribe::*;
 pub use self::try_connect::*;